moved canvas into its own file
